v 2.1 details fetching by id
